requirements.txt
beautifulsoup4>=4.12.3,<5.0
feedparser>=6.0.10,<7.0
python-telegram-bot>=20.7,<21.0
